requirements.txt
# Python requirements for Model Comparison CLI Tool
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
transformers>=4.40.0
torch>=2.1.0
